Break header down into composite parts.
